requirements.txt
django-storages[azure]
django
django-mssql-backend
django_widget_tweaks
pyodbc